Modified a lot of source code, on kinematics and CMake list.
I think we are finally in stable version. We still have to do a little bit of testing to be sure. Also I realized we don't need the calculate kinematics function for now. But it will be added later on.

+++ kinematics/spot_micro_matrix.cc
@@ -34,7 +34,7 @@ double get_determinant(const vector<vector<double>> vect) {
     if(vect.size() != vect[0].size()) {
         throw runtime_error("Matrix is not quadratic");
     } 
-    int dimension = vect.size();
+    size_t dimension = vect.size();
 
     if(dimension == 0) {
         return 1;
